Expose cli as library methods
