--- requirements.txt
@@ -1,3 +1,0 @@
---index-url https://pypi.python.org/simple/
-
--e .
